requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
torch==2.0.1
torchvision==0.15.2
Pillow==10.0.0